Rename REST folder and update namespaces
