+++ src/dotnet/CoreClient/Clients/RESTClients/CompletionRestClient.cs
@@ -1,13 +1,13 @@
-﻿using Azure.Core;
+﻿using System.Text;
+using System.Text.Json;
+using Azure.Core;
 using FoundationaLLM.Client.Core.Interfaces;
 using FoundationaLLM.Common.Models.Chat;
 using FoundationaLLM.Common.Models.Orchestration;
 using FoundationaLLM.Common.Models.ResourceProviders;
 using FoundationaLLM.Common.Models.ResourceProviders.Agent;
-using System.Text;
-using System.Text.Json;
 
-namespace FoundationaLLM.Client.Core.Clients.Rest
+namespace FoundationaLLM.Client.Core.Clients.RESTClients
 {
     /// <summary>
     /// Provides methods to manage calls to the Core API's orchestration endpoints.
