+++ src/dotnet/ManagementClient/Clients/RESTClients/IdentityRESTClient.cs
@@ -1,12 +1,12 @@
-﻿using Azure.Core;
+﻿using System.Text;
+using System.Text.Json;
+using Azure.Core;
 using FoundationaLLM.Client.Management.Interfaces;
 using FoundationaLLM.Common.Models.Authentication;
 using FoundationaLLM.Common.Settings;
 using Microsoft.Graph.Models;
-using System.Text;
-using System.Text.Json;
 
-namespace FoundationaLLM.Client.Management.Clients.Rest
+namespace FoundationaLLM.Client.Management.Clients.RESTClients
 {
     internal class IdentityRESTClient : ManagementRESTClientBase, IIdentityRESTClient
     {
