+++ src/dotnet/ManagementClient/Clients/RESTClients/ManagementRESTClientBase.cs
@@ -1,11 +1,11 @@
-﻿using Azure.Core;
+﻿using System.Text.Json;
+using Azure.Core;
 using FoundationaLLM.Common.Constants;
+using FoundationaLLM.Common.Constants.Authentication;
 using FoundationaLLM.Common.Extensions;
 using FoundationaLLM.Common.Settings;
-using System.Text.Json;
-using FoundationaLLM.Common.Constants.Authentication;
 
-namespace FoundationaLLM.Client.Management.Clients.Rest
+namespace FoundationaLLM.Client.Management.Clients.RESTClients
 {
     internal class ManagementRESTClientBase(
         IHttpClientFactory httpClientFactory,
