+++ src/dotnet/ManagementClient/Clients/RESTClients/StatusRESTClient.cs
@@ -1,11 +1,10 @@
-﻿using Azure;
+﻿using System.Text.Json;
 using Azure.Core;
 using FoundationaLLM.Client.Management.Interfaces;
 using FoundationaLLM.Common.Models.Infrastructure;
 using FoundationaLLM.Common.Settings;
-using System.Text.Json;
 
-namespace FoundationaLLM.Client.Management.Clients.Rest
+namespace FoundationaLLM.Client.Management.Clients.RESTClients
 {
     internal class StatusRESTClient(
         IHttpClientFactory httpClientFactory,
